# OpenApiConsultationsProject.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8081","description":"Generated server url"}],"paths":{"/api/medical_office/consultations":{"get":{"tags":["consultations-controller"],"summary":"Get all/searched consultation resources","operationId":"getAllConsultations","parameters":[{"name":"page","in":"query","required":false,"schema":{"minimum":0,"type":"integer","format":"int32"}},{"name":"itemsPerPage","in":"query","required":false,"schema":{"minimum":1,"type":"integer","format":"int32"}},{"name":"idPatient","in":"query","required":false,"schema":{"type":"string"}},{"name":"idDoctor","in":"query","required":false,"schema":{"type":"integer","format":"int32"}},{"name":"diagnostic","in":"query","required":false,"schema":{"type":"string"}},{"name":"appointmentTime","in":"query","required":false,"schema":{"type":"string","format":"date-time"}}],"responses":{"400":{"description":"Bad Request - Invalid syntax for query params","content":{}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"type":"string"}}}},"422":{"description":"Unprocessable Entity - Invalid values for query params","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnprocessableContentException"}}}},"406":{"description":"Not Acceptable - Requested representation not acceptable","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotAcceptableException"}}}},"404":{"description":"Not Found - Resource not found !","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotFoundException"}}}},"200":{"description":"Successful operation - Found all/searched resources","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ConsultationsCollection"}}}}}}},"post":{"tags":["consultations-controller"],"summary":"Create consultation resource  with POST because _id is autogenerated.","operationId":"createConsultation","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConsultationsCollection"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"409":{"description":"Conflict - Resource already exists","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConflictException"}}}},"422":{"description":"Unprocessable Entity - Invalid content","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnprocessableContentException"}}}},"406":{"description":"Not Acceptable - Requested representation not acceptable","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotAcceptableException"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}},"201":{"description":"Resource created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConsultationsCollection"}}}}}}},"/api/medical_office/consultations/consultation/{_id}/investigation":{"post":{"tags":["consultations-controller"],"summary":"Create investigation resource  with POST because _id is autogenerated.","operationId":"createInvestigation","parameters":[{"name":"_id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConsultationsCollection"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"409":{"description":"Conflict - Resource already exists","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConflictException"}}}},"422":{"description":"Unprocessable Entity - Invalid content","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnprocessableContentException"}}}},"406":{"description":"Not Acceptable - Requested representation not acceptable","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotAcceptableException"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}},"201":{"description":"Resource created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConsultationsCollection"}}}}}}},"/api/medical_office/consultations/{objectId}":{"patch":{"tags":["consultations-controller"],"summary":"Update consultation resource.","operationId":"updateConsultation","parameters":[{"name":"objectId","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConsultationsCollection"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"409":{"description":"Conflict - Requested representation conflicts with current state","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConflictException"}}}},"422":{"description":"Unprocessable Entity - Invalid content","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnprocessableContentException"}}}},"406":{"description":"Not Acceptable - Invalid request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotAcceptableException"}}}},"404":{"description":"Not Found - Resource not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotFoundException"}}}},"204":{"description":"Resource updated successfully"}}}},"/api/medical_office/consultations/{_id}":{"get":{"tags":["consultations-controller"],"summary":"Get consultation resource by _id field.","operationId":"getConsultationById","parameters":[{"name":"_id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"type":"string"}}}},"422":{"description":"Unprocessable Entity","content":{"*/*":{"schema":{"type":"string"}}}},"406":{"description":"Not Acceptable","content":{"*/*":{"schema":{"type":"string"}}}},"404":{"description":"Resource not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotFoundException"}}}},"200":{"description":"Resource found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConsultationsCollection"}}}}}},"delete":{"tags":["consultations-controller"],"summary":"Delete consultation resource.","operationId":"deleteConsultation","parameters":[{"name":"_id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"409":{"description":"Conflict","content":{"*/*":{"schema":{"type":"string"}}}},"422":{"description":"Unprocessable Entity","content":{"*/*":{"schema":{"type":"string"}}}},"406":{"description":"Not Acceptable - Invalid request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotAcceptableException"}}}},"404":{"description":"Not Found - Resource not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotFoundException"}}}},"200":{"description":"Resource deleted successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConsultationsCollection"}}}}}}}},"components":{"schemas":{"ConsultationsCollection":{"type":"object","properties":{"get_id":{"type":"string"},"idPatient":{"type":"string"},"idDoctor":{"type":"integer","format":"int32"},"appointmentTime":{"type":"string","format":"date-time"},"diagnostic":{"type":"string"},"investigations":{"type":"array","items":{"$ref":"#/components/schemas/InvestigationsCollection"}}}},"InvestigationsCollection":{"type":"object","properties":{"get_id":{"type":"string"},"naming":{"type":"string"},"processingHours":{"type":"integer","format":"int32"},"result":{"type":"string"}}},"NotAcceptableException":{"type":"object","properties":{"cause":{"type":"object","properties":{"stackTrace":{"type":"array","items":{"type":"object","properties":{"classLoaderName":{"type":"string"},"moduleName":{"type":"string"},"moduleVersion":{"type":"string"},"methodName":{"type":"string"},"fileName":{"type":"string"},"lineNumber":{"type":"integer","format":"int32"},"nativeMethod":{"type":"boolean"},"className":{"type":"string"}}}},"message":{"type":"string"},"localizedMessage":{"type":"string"}}},"stackTrace":{"type":"array","items":{"type":"object","properties":{"classLoaderName":{"type":"string"},"moduleName":{"type":"string"},"moduleVersion":{"type":"string"},"methodName":{"type":"string"},"fileName":{"type":"string"},"lineNumber":{"type":"integer","format":"int32"},"nativeMethod":{"type":"boolean"},"className":{"type":"string"}}}},"message":{"type":"string"},"suppressed":{"type":"array","items":{"type":"object","properties":{"stackTrace":{"type":"array","items":{"type":"object","properties":{"classLoaderName":{"type":"string"},"moduleName":{"type":"string"},"moduleVersion":{"type":"string"},"methodName":{"type":"string"},"fileName":{"type":"string"},"lineNumber":{"type":"integer","format":"int32"},"nativeMethod":{"type":"boolean"},"className":{"type":"string"}}}},"message":{"type":"string"},"localizedMessage":{"type":"string"}}}},"localizedMessage":{"type":"string"}}},"ConflictException":{"type":"object","properties":{"cause":{"type":"object","properties":{"stackTrace":{"type":"array","items":{"type":"object","properties":{"classLoaderName":{"type":"string"},"moduleName":{"type":"string"},"moduleVersion":{"type":"string"},"methodName":{"type":"string"},"fileName":{"type":"string"},"lineNumber":{"type":"integer","format":"int32"},"nativeMethod":{"type":"boolean"},"className":{"type":"string"}}}},"message":{"type":"string"},"localizedMessage":{"type":"string"}}},"stackTrace":{"type":"array","items":{"type":"object","properties":{"classLoaderName":{"type":"string"},"moduleName":{"type":"string"},"moduleVersion":{"type":"string"},"methodName":{"type":"string"},"fileName":{"type":"string"},"lineNumber":{"type":"integer","format":"int32"},"nativeMethod":{"type":"boolean"},"className":{"type":"string"}}}},"message":{"type":"string"},"suppressed":{"type":"array","items":{"type":"object","properties":{"stackTrace":{"type":"array","items":{"type":"object","properties":{"classLoaderName":{"type":"string"},"moduleName":{"type":"string"},"moduleVersion":{"type":"string"},"methodName":{"type":"string"},"fileName":{"type":"string"},"lineNumber":{"type":"integer","format":"int32"},"nativeMethod":{"type":"boolean"},"className":{"type":"string"}}}},"message":{"type":"string"},"localizedMessage":{"type":"string"}}}},"localizedMessage":{"type":"string"}}},"UnprocessableContentException":{"type":"object","properties":{"cause":{"type":"object","properties":{"stackTrace":{"type":"array","items":{"type":"object","properties":{"classLoaderName":{"type":"string"},"moduleName":{"type":"string"},"moduleVersion":{"type":"string"},"methodName":{"type":"string"},"fileName":{"type":"string"},"lineNumber":{"type":"integer","format":"int32"},"nativeMethod":{"type":"boolean"},"className":{"type":"string"}}}},"message":{"type":"string"},"localizedMessage":{"type":"string"}}},"stackTrace":{"type":"array","items":{"type":"object","properties":{"classLoaderName":{"type":"string"},"moduleName":{"type":"string"},"moduleVersion":{"type":"string"},"methodName":{"type":"string"},"fileName":{"type":"string"},"lineNumber":{"type":"integer","format":"int32"},"nativeMethod":{"type":"boolean"},"className":{"type":"string"}}}},"message":{"type":"string"},"suppressed":{"type":"array","items":{"type":"object","properties":{"stackTrace":{"type":"array","items":{"type":"object","properties":{"classLoaderName":{"type":"string"},"moduleName":{"type":"string"},"moduleVersion":{"type":"string"},"methodName":{"type":"string"},"fileName":{"type":"string"},"lineNumber":{"type":"integer","format":"int32"},"nativeMethod":{"type":"boolean"},"className":{"type":"string"}}}},"message":{"type":"string"},"localizedMessage":{"type":"string"}}}},"localizedMessage":{"type":"string"}}},"NotFoundException":{"type":"object","properties":{"cause":{"type":"object","properties":{"stackTrace":{"type":"array","items":{"type":"object","properties":{"classLoaderName":{"type":"string"},"moduleName":{"type":"string"},"moduleVersion":{"type":"string"},"methodName":{"type":"string"},"fileName":{"type":"string"},"lineNumber":{"type":"integer","format":"int32"},"nativeMethod":{"type":"boolean"},"className":{"type":"string"}}}},"message":{"type":"string"},"localizedMessage":{"type":"string"}}},"stackTrace":{"type":"array","items":{"type":"object","properties":{"classLoaderName":{"type":"string"},"moduleName":{"type":"string"},"moduleVersion":{"type":"string"},"methodName":{"type":"string"},"fileName":{"type":"string"},"lineNumber":{"type":"integer","format":"int32"},"nativeMethod":{"type":"boolean"},"className":{"type":"string"}}}},"message":{"type":"string"},"suppressed":{"type":"array","items":{"type":"object","properties":{"stackTrace":{"type":"array","items":{"type":"object","properties":{"classLoaderName":{"type":"string"},"moduleName":{"type":"string"},"moduleVersion":{"type":"string"},"methodName":{"type":"string"},"fileName":{"type":"string"},"lineNumber":{"type":"integer","format":"int32"},"nativeMethod":{"type":"boolean"},"className":{"type":"string"}}}},"message":{"type":"string"},"localizedMessage":{"type":"string"}}}},"localizedMessage":{"type":"string"}}}}}}